requirements.txt
python-dotenv
langchain-community
langchain-openai
langchain-tavily
langchain-chroma
transformers
gradio
notebook
ipywidgets
torch
langchain 
tiktoken 
langchainhub 
chromadb 
langgraph
chardet
python-docx

